refactors app.js
